+++ src/apps/plutus/plutus.module.ts
@@ -9,7 +9,7 @@ import { ArbitrumPlutusPlsDpxTokenFetcher } from './arbitrum/plutus.pls-dpx.toke
 import { ArbitrumPlutusPlsFarmContractPositionFetcher } from './arbitrum/plutus.pls-farm.contract-position-fetcher';
 import { ArbitrumPlutusPlsJonesFarmContractPositionFetcher } from './arbitrum/plutus.pls-jones-farm.contract-position-fetcher';
 import { ArbitrumPlutusPlsJonesTokenFetcher } from './arbitrum/plutus.pls-jones.token-fetcher';
-import { ArbitrumPlutusTgeContractPositionFetcher } from './arbitrum/plutus.tge.contract-position-fetcher';
+import { ArbitrumPlutusTgeClaimableContractPositionFetcher } from './arbitrum/plutus.tge-claimable.contract-position-fetcher';
 import { PlutusContractFactory } from './contracts';
 import { PlutusAppDefinition, PLUTUS_DEFINITION } from './plutus.definition';
 
@@ -24,7 +24,7 @@ import { PlutusAppDefinition, PLUTUS_DEFINITION } from './plutus.definition';
     ArbitrumPlutusPlsFarmContractPositionFetcher,
     ArbitrumPlutusPlsDpxTokenFetcher,
     ArbitrumPlutusPlsJonesTokenFetcher,
-    ArbitrumPlutusTgeContractPositionFetcher,
+    ArbitrumPlutusTgeClaimableContractPositionFetcher,
     PlutusAppDefinition,
     PlutusContractFactory,
   ],